refactor: 함수, 파일명 수정 및 경로 변경
