--- src/utils/api/logout/route.ts
@@ -1,18 +1,0 @@
-import { NextResponse } from 'next/server'
-
-export async function POST() {
-  const response = NextResponse.json({ success: true })
-
-  // accessToken 쿠키 삭제
-  response.cookies.set('accessToken', '', {
-    path: '/',
-    expires: new Date(0),
-  })
-
-  response.cookies.set('accessToken1', '', {
-    path: '/',
-    expires: new Date(0),
-  })
-
-  return response
-}
